pyproject.toml
[project]
name = "movie-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "fastmcp>=3.1.1",
    "openai>=2.30.0",
    "pydantic>=2.12.5",
    "pydantic-ai-slim[openai]>=1.73.0",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "ruff>=0.15.8",
]

[tool.ruff.lint]
extend-select = [
    "I", # Import Sorting
]
